# generated/api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

// Ref: #/components/schemas/Attribute
type Attribute struct {
	ID    string `json:"id"`
	Name  string `json:"name"`
	Icon  string `json:"icon"`
	Value string `json:"value"`
}

// GetID returns the value of ID.
func (s *Attribute) GetID() string {
	return s.ID
}

// GetName returns the value of Name.
func (s *Attribute) GetName() string {
	return s.Name
}

// GetIcon returns the value of Icon.
func (s *Attribute) GetIcon() string {
	return s.Icon
}

// GetValue returns the value of Value.
func (s *Attribute) GetValue() string {
	return s.Value
}

// SetID sets the value of ID.
func (s *Attribute) SetID(val string) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *Attribute) SetName(val string) {
	s.Name = val
}

// SetIcon sets the value of Icon.
func (s *Attribute) SetIcon(val string) {
	s.Icon = val
}

// SetValue sets the value of Value.
func (s *Attribute) SetValue(val string) {
	s.Value = val
}

// NewOptInt returns new OptInt with value set to v.
func NewOptInt(v int) OptInt {
	return OptInt{
		Value: v,
		Set:   true,
	}
}

// OptInt is optional int.
type OptInt struct {
	Value int
	Set   bool
}

// IsSet returns true if OptInt was set.
func (o OptInt) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptInt) Reset() {
	var v int
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptInt) SetTo(v int) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptInt) Get() (v int, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptInt) Or(d int) int {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/Resource
type Resource struct {
	Service     string      `json:"service"`
	Type        string      `json:"type"`
	ID          string      `json:"id"`
	Name        string      `json:"name"`
	Icon        string      `json:"icon"`
	Description OptString   `json:"description"`
	URL         OptString   `json:"url"`
	Attributes  []Attribute `json:"attributes"`
}

// GetService returns the value of Service.
func (s *Resource) GetService() string {
	return s.Service
}

// GetType returns the value of Type.
func (s *Resource) GetType() string {
	return s.Type
}

// GetID returns the value of ID.
func (s *Resource) GetID() string {
	return s.ID
}

// GetName returns the value of Name.
func (s *Resource) GetName() string {
	return s.Name
}

// GetIcon returns the value of Icon.
func (s *Resource) GetIcon() string {
	return s.Icon
}

// GetDescription returns the value of Description.
func (s *Resource) GetDescription() OptString {
	return s.Description
}

// GetURL returns the value of URL.
func (s *Resource) GetURL() OptString {
	return s.URL
}

// GetAttributes returns the value of Attributes.
func (s *Resource) GetAttributes() []Attribute {
	return s.Attributes
}

// SetService sets the value of Service.
func (s *Resource) SetService(val string) {
	s.Service = val
}

// SetType sets the value of Type.
func (s *Resource) SetType(val string) {
	s.Type = val
}

// SetID sets the value of ID.
func (s *Resource) SetID(val string) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *Resource) SetName(val string) {
	s.Name = val
}

// SetIcon sets the value of Icon.
func (s *Resource) SetIcon(val string) {
	s.Icon = val
}

// SetDescription sets the value of Description.
func (s *Resource) SetDescription(val OptString) {
	s.Description = val
}

// SetURL sets the value of URL.
func (s *Resource) SetURL(val OptString) {
	s.URL = val
}

// SetAttributes sets the value of Attributes.
func (s *Resource) SetAttributes(val []Attribute) {
	s.Attributes = val
}

// Ref: #/components/schemas/ResourceListResponse
type ResourceListResponse struct {
	Resources []Resource `json:"resources"`
}

// GetResources returns the value of Resources.
func (s *ResourceListResponse) GetResources() []Resource {
	return s.Resources
}

// SetResources sets the value of Resources.
func (s *ResourceListResponse) SetResources(val []Resource) {
	s.Resources = val
}

// Ref: #/components/schemas/ResourceType
type ResourceType struct {
	ID                 string   `json:"id"`
	Name               string   `json:"name"`
	EnrichmentPatterns []string `json:"enrichment_patterns"`
	Attributes         []string `json:"attributes"`
}

// GetID returns the value of ID.
func (s *ResourceType) GetID() string {
	return s.ID
}

// GetName returns the value of Name.
func (s *ResourceType) GetName() string {
	return s.Name
}

// GetEnrichmentPatterns returns the value of EnrichmentPatterns.
func (s *ResourceType) GetEnrichmentPatterns() []string {
	return s.EnrichmentPatterns
}

// GetAttributes returns the value of Attributes.
func (s *ResourceType) GetAttributes() []string {
	return s.Attributes
}

// SetID sets the value of ID.
func (s *ResourceType) SetID(val string) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *ResourceType) SetName(val string) {
	s.Name = val
}

// SetEnrichmentPatterns sets the value of EnrichmentPatterns.
func (s *ResourceType) SetEnrichmentPatterns(val []string) {
	s.EnrichmentPatterns = val
}

// SetAttributes sets the value of Attributes.
func (s *ResourceType) SetAttributes(val []string) {
	s.Attributes = val
}

// Ref: #/components/schemas/Service
type Service struct {
	ID            string         `json:"id"`
	Type          string         `json:"type"`
	ResourceTypes []ResourceType `json:"resource_types"`
}

// GetID returns the value of ID.
func (s *Service) GetID() string {
	return s.ID
}

// GetType returns the value of Type.
func (s *Service) GetType() string {
	return s.Type
}

// GetResourceTypes returns the value of ResourceTypes.
func (s *Service) GetResourceTypes() []ResourceType {
	return s.ResourceTypes
}

// SetID sets the value of ID.
func (s *Service) SetID(val string) {
	s.ID = val
}

// SetType sets the value of Type.
func (s *Service) SetType(val string) {
	s.Type = val
}

// SetResourceTypes sets the value of ResourceTypes.
func (s *Service) SetResourceTypes(val []ResourceType) {
	s.ResourceTypes = val
}

// Ref: #/components/schemas/ServiceListResponse
type ServiceListResponse struct {
	Services []Service `json:"services"`
}

// GetServices returns the value of Services.
func (s *ServiceListResponse) GetServices() []Service {
	return s.Services
}

// SetServices sets the value of Services.
func (s *ServiceListResponse) SetServices(val []Service) {
	s.Services = val
}